requirements.txt
pandas
scikit-learn
mysql-connector-python
flask
flask-cors
numpy
joblib